Add filter hook to modify translation array
Adds the `wpmlrestapi_get_translation` filter to allow for modification
of the translation array during `get_translations`. The filter is
applied once for each translation. The entire array is passed in along
with the post and the language.

Rather than change the href filter or the existing logic, this new
filter allows for a clean entrypoint to update the translation.

Use case example: site uses a permalink manager plugin that returns a
fully localized URL to the content. Using the existing
`wpmlrestapi_translations_href` filter passes the URL through the
rest of the function where it is incorrectly modified.